Revert "atualizando variavel projeto 2"
This reverts commit 6e060d426d469e290596838224c32703a73f6830.

+++ desafio_2/src/App.java
@@ -22,9 +22,9 @@ public class App {
         
         
         for (InteirosDinheiro dinheiro : listaInteirosDinheiro) {
-        	int quantidade;
+        	double quantidade;
         	
-        	quantidade = (int) (valorTotal/dinheiro.getValor());
+        	quantidade = valorTotal/dinheiro.getValor();
         	valorTotal %= dinheiro.getValor();	
         	System.out.println((int) quantidade +" "+ dinheiro.getTipo()+ "(s)(s) de R$ "+dinheiro.getNota());
 		}
